Refacto app.py
